better idle connection timeout handling
Previously, the idle timeout was started from the time the connection
was pulled out of the pool, so the idle timeout would start running
while the connection was in use for a query.

In practice, this was not ideal if you have queries or transactions that
span longer than the idle timeout.

This switches behavior to start the idle timeout from the point it's
returned to the pool, and the current idle timeout stopped when pulled
from the pool.

This should allow a timeout as aggressive as 1ms without any issues.

Fixes #15